Rename Endpoint -> RPC
