+++ api/src/api.ts
@@ -1,4 +1,4 @@
-import type { RequestOf, ResponseOf, RPC } from "./endpoint";
+import type { RequestOf, ResponseOf, RPC } from "./rpc";
 
 export interface Api {
   request<E extends RPC<unknown, unknown>>(
