+++ api/src/local.ts
@@ -1,6 +1,6 @@
 import type { CollectedTypes, ValueType } from "@previewjs/type-analyzer";
-import type { RPC } from "./endpoint";
 import type { PersistedState } from "./persisted-state";
+import type { RPC } from "./rpc";
 
 export const GetInfo: RPC<
   void,
